+++ tests/tracing/clients/test_tracking.py
@@ -1,10 +1,10 @@
-from mlflow.tracing.clients import InMemoryTraceClient
+from mlflow.tracing.clients import InMemoryTraceClientWithTracking
 
 
-def test_log_and_get_trace(monkeypatch, create_trace):
+def test_log_and_get_trace(monkeypatch, create_trace, mock_tracking_service_client):
     monkeypatch.setenv("MLFLOW_TRACING_CLIENT_BUFFER_SIZE", "3")
 
-    client = InMemoryTraceClient.get_instance()
+    client = InMemoryTraceClientWithTracking.get_instance()
     traces = client.get_traces()
     assert len(traces) == 0
